Traktor: Integrated latest dev to stable.
[git-p4: depot-paths = "//depot/traktor/": change = 16907]

+++ code/Core/Thread/ThreadLocal.h
@@ -26,7 +26,7 @@ public:
 	void* get() const;
 
 private:
-	uint32_t m_handle;
+	uint64_t m_handle;
 };
 
 }
